refactor(test): P5-3 follow-up — extract mask + overflow extension
Continue test_view.cpp split. Extracts two coherent View paint/hit-test
surfaces into a sibling TU:

  1. pulp #1148 slice (a) — symmetric overflow:visible hit-test
     extension (lines 1150-1231, 82 lines). PR #1297 added a 500px
     extension so absolutely-positioned popovers / dropdowns /
     menus protrude past their parent and still receive pointer
     hits. View::hit_test extends LEFT / RIGHT / UP / DOWN; does
     NOT extend past 500px.

  2. pulp #1737 / #1515 — CSS mask-image + mask-size paint slice
     (lines 1330-1406, 77 lines). View::paint_all routes through
     save_layer_with_mask when mask-image is set; does NOT route
     through when empty or 'none'.

* test_view.cpp: 1,466 → 1,307 lines (-159)
* New target pulp-test-view-mask-overflow: 190 lines, 8 TEST_CASEs,
  14 assertions.
